automation/api/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-filter>=25.0,<26.0
drf-spectacular>=0.28,<1.0
gunicorn>=23.0,<24.0
psycopg2-binary>=2.9,<3.0
